pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.14",
    "discord-py>=2.5.2",
    "flask>=3.1.1",
    "python-dotenv>=1.1.1",
]
